Register post

+++ Spark.Core/IntRepository/IUserRepository.cs
@@ -9,7 +9,7 @@ namespace Spark.Core.IntRepository
 {
     public interface IUserRepository:IRepository<User>
     {
-        Task<IEnumerable<User>> GetUserByGenderAndAge(string gender,short minAge,short maxAge);
+        Task<IEnumerable<User>> GetUserByGenderAndAge(string gender,int minAge,int maxAge);
 
         double CalculateDistance(Guid user1Id, Guid user2Id);
