+++ Spark.Core/Models/User.cs
@@ -11,7 +11,7 @@ namespace Spark.Core.Models
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string LastName { get; set; }
-        public short Age { get; set; }
+        public int Age { get; set; }
         public string Gender { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
